koniec testy dzialają, dodaję je również

+++ src/helpers.cpp
@@ -3,7 +3,7 @@
 //
 #include <random>
 #include "helpers.hpp"
-ProbabilityGenerator probability_generator=[](){return rand()/ RAND_MAX;};
+ProbabilityGenerator probability_generator=[](){return (double)rand()/ RAND_MAX;};
